requirements.txt
numpy
matplotlib
pyngrok
bing_image_downloader
scikit_image
scikit-learn
Pillow
streamlit
joblib